requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptoapis==1.7.0
Django==4.2.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.1.12
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
PyJWT==2.7.0
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
webencodings==0.5.1
